Corrigindo tipo do parametro da funcao que imprime o candidato

+++ 04_TAD_simples/TAD_02/Respostas/CaioCostaLopes/candidato.h
@@ -85,6 +85,6 @@ float CalculaPercentualVotos(tCandidato candidatos, int totalVotos);
  * @param candidato Candidato que terá seus dados impressos;
  * @param percentualVotos Percentual (porcentagem) do total de votos que o candidato obteve na eleição;
  */
-void ImprimeCandidato(tCandidato candidato, int percentualVotos);
+void ImprimeCandidato(tCandidato candidato, float percentualVotos);
 
 #endif
